fix: detect correct package manager
Fixes #14

This issue stemmed from `app-root-path` detecting from the directory of
the executable rather than `process.cwd`. To resolve this issue, I
replaced `app-root-path` with `pkg-dir` which detects the correct
package directory. Additionally, `pkg-dir` is the base for `pkg-up` so
we're now using similar tools for locating both the package directory
and `package.json` file.